requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-environ==0.8.1
django-filter==21.1
django-modeltranslation==0.17.3
djangorestframework==3.12.4
mysqlclient==2.1.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
